refactor(shared): split the validator, constant and event barrels by domain
Three files were growing without bound and every stream needed to append to
them, so they became the merge bottleneck for the whole programme. Each is now
a directory of per-domain modules with index.ts kept as a pure re-export
barrel, so a stream owns one small file instead of contending on one big one.

Purely mechanical: no export renamed, no schema or value changed, no import
site touched. Equivalence was checked by importing both barrels side by side
and comparing export keys, scalar values and every Zod schema serialised
through toJSONSchema.
