+++ packages/shared/src/constants/actor.ts
@@ -1,0 +1,2 @@
+export const ACTOR_TYPES = ['user', 'integration', 'agent', 'system'] as const;
+export type ActorType = (typeof ACTOR_TYPES)[number];
