+++ packages/shared/src/constants/doc.ts
@@ -1,0 +1,2 @@
+export const DOC_VISIBILITIES = ['workspace', 'link', 'public'] as const;
+export type DocVisibility = (typeof DOC_VISIBILITIES)[number];
