+++ packages/shared/src/constants/estimate.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_ESTIMATE_SCALE = [0, 1, 2, 3, 5, 8, 13] as const;
