+++ packages/shared/src/constants/ordering.ts
@@ -1,0 +1,1 @@
+export const SORT_ORDER_STEP = 1024;
