+++ packages/shared/src/constants/pattern.ts
@@ -1,0 +1,3 @@
+export const IDENTIFIER_PATTERN = /^[A-Z][A-Z0-9]{1,5}$/;
+
+export const SLUG_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
