+++ packages/shared/src/constants/project.ts
@@ -1,0 +1,11 @@
+export const PROJECT_HEALTHS = ['on_track', 'at_risk', 'off_track', 'no_update'] as const;
+export type ProjectHealth = (typeof PROJECT_HEALTHS)[number];
+
+export const PROJECT_STATUSES = [
+  'backlog',
+  'planned',
+  'in_progress',
+  'completed',
+  'canceled',
+] as const;
+export type ProjectStatus = (typeof PROJECT_STATUSES)[number];
