+++ packages/shared/src/events/channel.ts
@@ -1,0 +1,4 @@
+export const REDIS_DELTA_CHANNEL = 'orbit:delta';
+export const REDIS_PRESENCE_CHANNEL = 'orbit:presence';
+
+export const DELTA_BATCH_WINDOW_MS = 50;
